replace in file if the file exists
Currently when you try to install Breeze with Blade and you don't have the welcome.blade.php file it fails because it is not able to find the welcome.blade.php

+++ src/Console/InstallCommand.php
@@ -247,7 +247,9 @@ class InstallCommand extends Command implements PromptsForMissingInput
      */
     protected function replaceInFile($search, $replace, $path)
     {
-        file_put_contents($path, str_replace($search, $replace, file_get_contents($path)));
+        if (File::exists($path)) {
+            file_put_contents($path, str_replace($search, $replace, file_get_contents($path)));
+        }
     }
 
     /**
